Get MSBuild desktop and Core working at once
