+++ src/CodeGeneration.Roslyn.Tasks/App.config
@@ -1,0 +1,10 @@
+﻿<configuration>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="System.Collections.Immutable" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+        <bindingRedirect oldVersion="1.0.0.0-1.2.0.0" newVersion="1.1.37.0" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
+</configuration>
