+++ src/CodeGeneration.Roslyn.Tasks/CodeGeneration.Roslyn.Tasks.csproj
@@ -45,21 +45,19 @@
     </ProjectReference>
   </ItemDefinitionGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\CodeGeneration.Roslyn\CodeGeneration.Roslyn.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="1.3.2" />
-  </ItemGroup>
-
   <ItemGroup Condition="'$(TargetFramework)' == 'net452'">
     <PackageReference Include="Microsoft.Build.Tasks.Core" Version="14.3" Condition=" '$(TargetFramework)' == 'net452' " />
+    <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
 
   <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp1.0'">
     <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
     <PackageReference Include="Microsoft.Build.Tasks.Core" Version="15.1.0-preview-000458-02" />
+  </ItemGroup>
+
+  <ItemGroup>
+    
+    <ProjectReference Include="..\CodeGeneration.Roslyn.Tasks.Helper\CodeGeneration.Roslyn.Tasks.Helper.csproj" />
   </ItemGroup>
   
 </Project>
