+++ src/CodeGeneration.Roslyn.Tasks/GenerateCodeFromAttributes.cs
@@ -16,14 +16,15 @@ namespace CodeGeneration.Roslyn.Tasks
     using System.Threading;
     using Microsoft.Build.Framework;
     using Microsoft.Build.Utilities;
-    using Microsoft.CodeAnalysis;
-    using Microsoft.CodeAnalysis.CSharp;
-    using Microsoft.CodeAnalysis.Text;
     using Task = System.Threading.Tasks.Task;
 
     public class GenerateCodeFromAttributes : Microsoft.Build.Utilities.Task, ICancelableTask
     {
         private readonly CancellationTokenSource cts = new CancellationTokenSource();
+
+#if NETCOREAPP1_0
+        private TaskLoadContext taskLoadContext;
+#endif
 
         [Required]
         public ITaskItem[] Compile { get; set; }
@@ -54,15 +55,23 @@ namespace CodeGeneration.Roslyn.Tasks
             // Run under our own AppDomain so we can control the version of Roslyn we load.
             var appDomainSetup = new AppDomainSetup();
             appDomainSetup.ApplicationBase = Path.GetDirectoryName(this.GetType().Assembly.Location);
-            appDomainSetup.DisallowBindingRedirects = true; // We want the version of Roslyn we compiled against.
+            appDomainSetup.ConfigurationFile = Assembly.GetExecutingAssembly().Location + ".config";
             var appDomain = AppDomain.CreateDomain("codegen", AppDomain.CurrentDomain.Evidence, appDomainSetup);
+            AppDomain.CurrentDomain.AssemblyResolve += this.CurrentDomain_AssemblyResolve;
+#else
+
 #endif
             try
             {
+                var helperAssemblyName = new AssemblyName(typeof(GenerateCodeFromAttributes).GetTypeInfo().Assembly.GetName().FullName.Replace(ThisAssembly.AssemblyName, ThisAssembly.AssemblyName + ".Helper"));
+                const string helperTypeName = "CodeGeneration.Roslyn.Tasks.Helper";
 #if NET452
-                var helper = (Helper)appDomain.CreateInstanceAndUnwrap(Assembly.GetExecutingAssembly().FullName, typeof(Helper).FullName);
+                var helper = (Helper)appDomain.CreateInstanceAndUnwrap(helperAssemblyName.FullName, helperTypeName);
 #else
-                var helper = new Helper();
+                this.taskLoadContext = new TaskLoadContext(Path.GetDirectoryName(new Uri(typeof(GenerateCodeFromAttributes).GetTypeInfo().Assembly.CodeBase).LocalPath));
+                var helperAssembly = this.taskLoadContext.LoadFromAssemblyName(helperAssemblyName);
+                var helperTypeInContext = helperAssembly.GetType(helperTypeName);
+                dynamic helper = Activator.CreateInstance(helperTypeInContext, this.taskLoadContext);
 #endif
                 helper.Compile = this.Compile;
                 helper.TargetName = this.TargetName;
@@ -77,8 +86,8 @@ namespace CodeGeneration.Roslyn.Tasks
 
                     // Copy the contents of the output parameters into our own. Don't just copy the reference
                     // because we're going to unload the AppDomain.
-                    this.GeneratedCompile = helper.GeneratedCompile.Select(i => new TaskItem(i)).ToArray();
-                    this.AdditionalWrittenFiles = helper.AdditionalWrittenFiles.Select(i => new TaskItem(i)).ToArray();
+                    this.GeneratedCompile = ((IEnumerable<ITaskItem>)helper.GeneratedCompile).Select(i => new TaskItem(i)).ToArray();
+                    this.AdditionalWrittenFiles = ((IEnumerable<ITaskItem>)helper.AdditionalWrittenFiles).Select(i => new TaskItem(i)).ToArray();
 
                     return !this.Log.HasLoggedErrors;
                 }
@@ -91,292 +100,60 @@ namespace CodeGeneration.Roslyn.Tasks
             finally
             {
 #if NET452
+                AppDomain.CurrentDomain.AssemblyResolve -= this.CurrentDomain_AssemblyResolve;
                 AppDomain.Unload(appDomain);
 #endif
             }
         }
+
+#if NET452
+        private Assembly CurrentDomain_AssemblyResolve(object sender, ResolveEventArgs args)
+        {
+            try
+            {
+                return Assembly.Load(args.Name);
+            }
+            catch
+            {
+                return null;
+            }
+        }
+#endif
 
         public void Cancel()
         {
             this.cts.Cancel();
         }
 
-        private class Helper
-#if NET452
-            : MarshalByRefObject
-#endif
+#if NETCOREAPP1_0
+        private class TaskLoadContext : AssemblyLoadContext
         {
-            private readonly List<string> loadedAssemblies = new List<string>();
+            private readonly string assemblyLoadPath;
 
-            public CancellationToken CancellationToken { get; set; }
-
-            public ITaskItem[] Compile { get; set; }
-
-            public string TargetName { get; set; }
-
-            public ITaskItem[] ReferencePath { get; set; }
-
-            public ITaskItem[] GeneratorAssemblySearchPaths { get; set; }
-
-            public string IntermediateOutputDirectory { get; set; }
-
-            public ITaskItem[] GeneratedCompile { get; set; }
-
-            public ITaskItem[] AdditionalWrittenFiles { get; set; }
-
-            public TaskLoggingHelper Log { get; set; }
-
-            public void Execute()
+            internal TaskLoadContext(string assemblyLoadPath)
             {
-                Task.Run(async delegate
-                {
-#if NET452
-                    AppDomain.CurrentDomain.AssemblyResolve += (s, e) =>
-                    {
-                        return this.TryLoadAssembly(new AssemblyName(e.Name));
-                    };
-#else
-                    AssemblyLoadContext.Default.Resolving += (lc, an) =>
-                    {
-                        return this.TryLoadAssembly(an);
-                    };
-#endif
-
-                    var project = this.CreateProject();
-                    var outputFiles = new List<ITaskItem>();
-                    var writtenFiles = new List<ITaskItem>();
-
-                    string generatorAssemblyInputsFile = Path.Combine(this.IntermediateOutputDirectory, "CodeGeneration.Roslyn.InputAssemblies.txt");
-
-                    // For incremental build, we want to consider the input->output files as well as the assemblies involved in code generation.
-                    DateTime assembliesLastModified = GetLastModifiedAssemblyTime(generatorAssemblyInputsFile);
-
-                    var explicitIncludeList = new HashSet<string>(
-                        from item in this.Compile
-                        where string.Equals(item.GetMetadata("Generator"), $"MSBuild:{this.TargetName}", StringComparison.OrdinalIgnoreCase)
-                        select item.ItemSpec,
-                        StringComparer.OrdinalIgnoreCase);
-
-                    foreach (var inputDocument in project.Documents)
-                    {
-                        this.CancellationToken.ThrowIfCancellationRequested();
-
-                        // Skip over documents that aren't on the prescribed list of files to scan.
-                        if (!explicitIncludeList.Contains(inputDocument.Name))
-                        {
-                            continue;
-                        }
-
-                        string sourceHash = inputDocument.Name.GetHashCode().ToString("x", CultureInfo.InvariantCulture);
-                        string outputFilePath = Path.Combine(this.IntermediateOutputDirectory, Path.GetFileNameWithoutExtension(inputDocument.Name) + $".{sourceHash}.generated.cs");
-
-                        // Code generation is relatively fast, but it's not free.
-                        // And when we run the Simplifier.ReduceAsync it's dog slow.
-                        // So skip files that haven't changed since we last generated them.
-                        bool generated = false;
-                        DateTime outputLastModified = File.GetLastWriteTime(outputFilePath);
-                        if (File.GetLastWriteTime(inputDocument.Name) > outputLastModified || assembliesLastModified > outputLastModified)
-                        {
-                            this.Log.LogMessage(MessageImportance.Normal, "{0} -> {1}", inputDocument.Name, outputFilePath);
-
-                            var outputDocument = await DocumentTransform.TransformAsync(
-                                inputDocument,
-                                new ProgressLogger(this.Log, inputDocument.Name));
-
-                            // Only produce a new file if the generated document is not empty.
-                            var semanticModel = await outputDocument.GetSemanticModelAsync(this.CancellationToken);
-                            if (!CSharpDeclarationComputer.GetDeclarationsInSpan(semanticModel, TextSpan.FromBounds(0, semanticModel.SyntaxTree.Length), false, this.CancellationToken).IsEmpty)
-                            {
-                                var outputText = await outputDocument.GetTextAsync(this.CancellationToken);
-                                using (var outputFileStream = File.OpenWrite(outputFilePath))
-                                using (var outputWriter = new StreamWriter(outputFileStream))
-                                {
-                                    outputText.Write(outputWriter);
-
-                                    // Truncate any data that may be beyond this point if the file existed previously.
-                                    outputWriter.Flush();
-                                    outputFileStream.SetLength(outputFileStream.Position);
-                                }
-
-                                generated = true;
-                            }
-                        }
-                        else
-                        {
-                            generated = true;
-                        }
-
-                        if (generated)
-                        {
-                            var outputItem = new TaskItem(outputFilePath);
-                            outputFiles.Add(outputItem);
-                        }
-                    }
-
-                    this.SaveGeneratorAssemblyList(generatorAssemblyInputsFile);
-                    writtenFiles.Add(new TaskItem(generatorAssemblyInputsFile));
-
-                    this.GeneratedCompile = outputFiles.ToArray();
-                    this.AdditionalWrittenFiles = writtenFiles.ToArray();
-                }).GetAwaiter().GetResult();
+                this.assemblyLoadPath = assemblyLoadPath;
             }
 
-            private static DateTime GetLastModifiedAssemblyTime(string assemblyListPath)
+            protected override Assembly Load(AssemblyName assemblyName)
             {
-                if (!File.Exists(assemblyListPath))
+                if (assemblyName.Name.StartsWith("Microsoft.Build", StringComparison.OrdinalIgnoreCase) ||
+                    assemblyName.Name.StartsWith("System.", StringComparison.OrdinalIgnoreCase))
                 {
-                    return DateTime.MinValue;
+                    // MSBuild and System.* make up our exchange types. So don't load them in this LoadContext.
+                    // We need to inherit them from the default load context.
+                    return null;
                 }
 
-                var timestamps = (from path in File.ReadAllLines(assemblyListPath)
-                                  where File.Exists(path)
-                                  select File.GetLastWriteTime(path)).ToList();
-                return timestamps.Any() ? timestamps.Max() : DateTime.MinValue;
-            }
-
-            private void SaveGeneratorAssemblyList(string assemblyListPath)
-            {
-                // Union our current list with the one on disk, since our incremental code generation
-                // may have skipped some up-to-date files, resulting in fewer assemblies being loaded
-                // this time.
-                var assemblyPaths = new SortedSet<string>(StringComparer.OrdinalIgnoreCase);
-
-                if (File.Exists(assemblyListPath))
+                string assemblyPath = Path.Combine(this.assemblyLoadPath, assemblyName.Name) + ".dll";
+                if (File.Exists(assemblyPath))
                 {
-                    assemblyPaths.UnionWith(File.ReadAllLines(assemblyListPath));
-                }
-
-                assemblyPaths.UnionWith(this.loadedAssemblies);
-
-#if NET452
-                assemblyPaths.UnionWith(
-                    from a in AppDomain.CurrentDomain.GetAssemblies()
-                    where !a.IsDynamic
-                    select a.Location);
-#endif
-
-                File.WriteAllLines(
-                    assemblyListPath,
-                    assemblyPaths);
-            }
-
-            private static Assembly LoadAssemblyByFile(string path)
-            {
-#if NET452
-                return Assembly.LoadFile(path);
-#else
-                return AssemblyLoadContext.Default.LoadFromAssemblyPath(path);
-#endif
-            }
-
-            private Assembly TryLoadAssembly(AssemblyName assemblyName)
-            {
-                try
-                {
-                    var referencePath = this.ReferencePath.FirstOrDefault(rp => string.Equals(rp.GetMetadata("FileName"), assemblyName.Name, StringComparison.OrdinalIgnoreCase));
-                    if (referencePath != null)
-                    {
-                        string fullPath = referencePath.GetMetadata("FullPath");
-                        this.loadedAssemblies.Add(fullPath);
-                        return LoadAssemblyByFile(fullPath);
-                    }
-
-                    foreach (var searchPath in this.GeneratorAssemblySearchPaths)
-                    {
-                        string searchDir = searchPath.GetMetadata("FullPath");
-                        const string extension = ".dll";
-                        string fileName = Path.Combine(searchDir, assemblyName.Name + extension);
-                        if (File.Exists(fileName))
-                        {
-                            return LoadAssemblyByFile(fileName);
-                        }
-                    }
-                }
-                catch (BadImageFormatException)
-                {
+                    return LoadFromAssemblyPath(assemblyPath);
                 }
 
                 return null;
             }
-
-            private Project CreateProject()
-            {
-                var workspace = new AdhocWorkspace();
-                var project = workspace.CurrentSolution.AddProject("codegen", "codegen", "C#")
-                    .WithCompilationOptions(new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary))
-                    .WithMetadataReferences(this.ReferencePath.Select(p => MetadataReference.CreateFromFile(p.ItemSpec)));
-
-                foreach (var sourceFile in this.Compile)
-                {
-                    using (var stream = File.OpenRead(sourceFile.ItemSpec))
-                    {
-                        this.CancellationToken.ThrowIfCancellationRequested();
-                        var text = SourceText.From(stream);
-                        project = project.AddDocument(sourceFile.ItemSpec, text).Project;
-                    }
-                }
-
-                return project;
-            }
         }
-
-        private class ProgressLogger : IProgress<Diagnostic>
-        {
-            private readonly TaskLoggingHelper logger;
-            private readonly string inputFilename;
-
-            internal ProgressLogger(TaskLoggingHelper logger, string inputFilename)
-            {
-                this.logger = logger;
-                this.inputFilename = inputFilename;
-            }
-
-            public void Report(Diagnostic value)
-            {
-                var lineSpan = value.Location.GetLineSpan();
-                switch (value.Severity)
-                {
-                    case DiagnosticSeverity.Info:
-                        this.logger.LogMessage(
-                            value.Descriptor.Category,
-                            value.Descriptor.Id,
-                            value.Descriptor.HelpLinkUri,
-                            value.Location.SourceTree.FilePath,
-                            lineSpan.StartLinePosition.Line + 1,
-                            lineSpan.StartLinePosition.Character + 1,
-                            lineSpan.EndLinePosition.Line + 1,
-                            lineSpan.EndLinePosition.Character + 1,
-                            MessageImportance.Normal,
-                            value.GetMessage(CultureInfo.CurrentCulture));
-                        break;
-                    case DiagnosticSeverity.Warning:
-                        this.logger.LogWarning(
-                            value.Descriptor.Category,
-                            value.Descriptor.Id,
-                            value.Descriptor.HelpLinkUri,
-                            value.Location.SourceTree.FilePath,
-                            lineSpan.StartLinePosition.Line + 1,
-                            lineSpan.StartLinePosition.Character + 1,
-                            lineSpan.EndLinePosition.Line + 1,
-                            lineSpan.EndLinePosition.Character + 1,
-                            value.GetMessage(CultureInfo.CurrentCulture));
-                        break;
-                    case DiagnosticSeverity.Error:
-                        this.logger.LogError(
-                            value.Descriptor.Category,
-                            value.Descriptor.Id,
-                            value.Descriptor.HelpLinkUri,
-                            value.Location.SourceTree.FilePath,
-                            lineSpan.StartLinePosition.Line + 1,
-                            lineSpan.StartLinePosition.Character + 1,
-                            lineSpan.EndLinePosition.Line + 1,
-                            lineSpan.EndLinePosition.Character + 1,
-                            value.GetMessage(CultureInfo.CurrentCulture));
-                        break;
-                    default:
-                        break;
-                }
-            }
-        }
+#endif
     }
 }
