+++ src/CodeGeneration.Roslyn.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26217.2
+VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0BA25C7A-1A7A-43DA-BF47-202B54C75C9B}"
 	ProjectSection(SolutionItems) = preProject
@@ -26,6 +26,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CodeGeneration.Roslyn.Tests.Generators", "CodeGeneration.Roslyn.Tests.Generators\CodeGeneration.Roslyn.Tests.Generators.csproj", "{8B57E557-EF79-47E4-9396-135850304990}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CodeGeneration.Roslyn.Tests.RunBuild", "CodeGeneration.Roslyn.Tests.RunBuild\CodeGeneration.Roslyn.Tests.RunBuild.csproj", "{3162A2B3-5FEF-454C-AB50-A103883CC2F4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeGeneration.Roslyn.Tasks.Helper", "CodeGeneration.Roslyn.Tasks.Helper\CodeGeneration.Roslyn.Tasks.Helper.csproj", "{5C75670D-25A7-4707-B7F1-F9E523B69FC0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,6 +59,10 @@ Global
 		{3162A2B3-5FEF-454C-AB50-A103883CC2F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3162A2B3-5FEF-454C-AB50-A103883CC2F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3162A2B3-5FEF-454C-AB50-A103883CC2F4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C75670D-25A7-4707-B7F1-F9E523B69FC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C75670D-25A7-4707-B7F1-F9E523B69FC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C75670D-25A7-4707-B7F1-F9E523B69FC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C75670D-25A7-4707-B7F1-F9E523B69FC0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
